[05052022]
Major script update & relocate
