+++ src/js/aircraftSlot.json
