+++ src/js/fighter.json
