+++ src/js/skin_path.js
