+++ src/js/torpedoBomber.json
